Fixed check admin bugs

+++ controllers/modelsController.php
@@ -47,54 +47,55 @@
         }
 
         public function commandAddModel(){
-            $this->checkAdmin();
-            if ($_POST['p_selection']!=''
-            &&$_POST['m_name']!=''
-            &&$_POST['m_description']!=''
-            &&$_FILES['m_photo']!='') {
-                move_uploaded_file($_FILES['m_photo']['tmp_name'], "images/".$_FILES['m_photo']['name']);
-                $mod = array($_POST['p_selection'],
-                        $_POST['m_name'],
-                        $_POST['m_description'],
-                        $_FILES['m_photo']['name']);
-                $this->model->addModel($mod);
+            if($this->checkAdmin()){
+                if ($_POST['p_selection']!=''
+                &&$_POST['m_name']!=''
+                &&$_POST['m_description']!=''
+                &&$_FILES['m_photo']!='') {
+                    move_uploaded_file($_FILES['m_photo']['tmp_name'], "images/".$_FILES['m_photo']['name']);
+                    $mod = array($_POST['p_selection'],
+                            $_POST['m_name'],
+                            $_POST['m_description'],
+                            $_FILES['m_photo']['name']);
+                    $this->model->addModel($mod);
+                }
             }
             header("Location: ".URL);
         }
 
         public function commandEditModel($prodName, $modId, $modName, $mPhoto){
-            $this->checkAdmin();
-            if (($_POST['m_name']!='') && ($_POST['m_description']!='') && ($_FILES['m_photo']!='')) {
-                unlink("./././././images/".$mPhoto);
-                move_uploaded_file($_FILES['m_photo']['tmp_name'], "./images/".$_FILES['m_photo']['name']);
-                // First parameter is the actual direction of the file,
-                // second is its new direction. Notice that PHP uses only 
-                // one dot to come back in the directory while HTML uses
-                // two.
-                $editedModAndId = array($_POST['m_name'],
-                        $_POST['m_description'],
-                        $_FILES['m_photo']['name'],
-                        $modId);
-                        // Instead of sending the photo straight
-                        // we must send the link to it.
+            if($this->checkAdmin()){
+                if (($_POST['m_name']!='') && ($_POST['m_description']!='') && ($_FILES['m_photo']!='')) {
+                    unlink("./././././images/".$mPhoto);
+                    move_uploaded_file($_FILES['m_photo']['tmp_name'], "./images/".$_FILES['m_photo']['name']);
+                    // First parameter is the actual direction of the file,
+                    // second is its new direction. Notice that PHP uses only 
+                    // one dot to come back in the directory while HTML uses
+                    // two.
+                    $editedModAndId = array($_POST['m_name'],
+                            $_POST['m_description'],
+                            $_FILES['m_photo']['name'],
+                            $modId);
+                            // Instead of sending the photo straight
+                            // we must send the link to it.
 
-                // Here I need a verification of non repeated name.
-                // FAKE: you don't need it because modify options are
-                // only in that specific model.
-                $this->model->editModel($editedModAndId);
-                header("Location: ".URL.$prodName."/".$_POST['m_name']);
-                
+                    // Here I need a verification of non repeated name.
+                    // FAKE: you don't need it because modify options are
+                    // only in that specific model.
+                    $this->model->editModel($editedModAndId);
+                    header("Location: ".URL.$prodName."/".$_POST['m_name']);
+                    die();
+                }
             }
-            else{
-                header("Location: ".URL.$prodName."/".$modName);
-            }
+            header("Location: ".URL.$prodName."/".$modName);
         }
 
         public function commandDelModel($mId, $mPhoto){
-            $this->checkAdmin();
-            unlink("./././images/".$mPhoto);
-            // How to delete the photo?? 0.0
-            $this->model->delModel($mId);
+            if($this->checkAdmin()){
+                unlink("./././images/".$mPhoto);
+                // How to delete the photo?? 0.0
+                $this->model->delModel($mId);
+            }
             header("Location: ".URL);
         }
     }
